Move KnownFrameworkReference to a different type
